requirements.txt
akshare
pandas
numpy
google-generativeai
